blueprints/examples/EKS/infra/cdk/requirements.txt
aws-cdk-lib==2.51.1
boto3==1.24.77
constructs>=10.0.0,<11.0.0
